refactor abstraction of FormatProcessor
