+++ library/src/main/java/com/ibm/research/drl/dpt/processors/HL7FormatProcessor.java
@@ -26,7 +26,7 @@ import java.io.InputStreamReader;
 import java.util.Collections;
 import java.util.Set;
 
-public class HL7FormatProcessor extends AbstractFormatProcessor {
+public class HL7FormatProcessor extends FormatProcessor {
 
 
     @Override
